requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
dj-database-url==1.0.0
Django==5.2.7
django-environ==0.11.2
django-filter==25.2
djangorestframework==3.16.1
dotenv==0.9.9
gunicorn==20.1.0
idna==3.11
numpy==1.26.4
pandas==2.1.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
drf-nested-routers==0.95.0